feat: added ContractApiClient
