+++ src/api/DashboardApiClient.ts
@@ -10,22 +10,12 @@ import { ENV } from '../utils/env';
 import { WithDefaultsT, createClient } from './generated/b4f-dashboard/client';
 import { InstitutionResource } from './generated/b4f-dashboard/InstitutionResource';
 import { ProductsResource } from './generated/b4f-dashboard/ProductsResource';
-import { ContractTemplateResponseList } from './generated/b4f-dashboard/ContractTemplateResponseList';
-import { ContractTemplateFile } from './generated/b4f-dashboard/ContractTemplateFile';
-import { ContractTemplateUploadRequest } from './generated/b4f-dashboard/ContractTemplateUploadRequest';
 
 const MOCK_BEARER_TOKEN = storageTokenOps.read();
-
-const MOCK_BEARER_TOKEN_CONTRACTS = storageTokenOps.read();
 
 const withBearerAndPartyId: WithDefaultsT<'bearerAuth'> = (wrappedOperation) => (params: any) => wrappedOperation({
   ...params,
   bearerAuth: `Bearer ${MOCK_BEARER_TOKEN}`,
-});
-
-const withBearerAndPartyIdContracts: WithDefaultsT<'bearerAuth'> = (wrappedOperation) => (params: any) => wrappedOperation({
-  ...params,
-  bearerAuth: `Bearer ${MOCK_BEARER_TOKEN_CONTRACTS}`,
 });
 
 const apiClient = createClient({
@@ -33,13 +23,6 @@ const apiClient = createClient({
   basePath: '',
   fetchApi: buildFetchApi(ENV.API_TIMEOUT_MS.DASHBOARD),
   withDefaults: withBearerAndPartyId,
-});
-
-const apiClientContracts = createClient({
-  baseUrl: ENV.URL_API.API_DASHBOARD,
-  basePath: '',
-  fetchApi: buildFetchApi(ENV.API_TIMEOUT_MS.DASHBOARD),
-  withDefaults: withBearerAndPartyIdContracts,
 });
 
 const onRedirectToLogin = () =>
@@ -81,38 +64,5 @@ export const DashboardApi = {
       lang,
     });
     return extractResponse(result, 200, onRedirectToLogin);
-  },
-
-  listContractTemplates: async (
-    name?: string,
-    version?: string
-  ): Promise<ContractTemplateResponseList> => {
-    const result = await apiClientContracts.listContractTemplates({
-      name,
-      version,
-    });
-    return extractResponse(result, 200, onRedirectToLogin);
-  },
-
-  downloadContractTemplate: async (
-    contractId: string,
-    productId: string
-  ): Promise<ContractTemplateFile> => {
-    const result = await apiClientContracts.downloadContractTemplate({
-      contractId,
-      productId,
-    });
-    return extractResponse(result, 200, onRedirectToLogin);
-  },
-
-  postUploadContract: async (
-    productId: string,
-    contractTemplateUploadRequest: ContractTemplateUploadRequest
-  ): Promise<void> => {
-    const result = await apiClientContracts.postUploadContract({
-      productId,
-      body: contractTemplateUploadRequest,
-    });
-    return extractResponse(result, 201, onRedirectToLogin);
   },
 };
